chore: reshuffle folders.
