+++ instrumentation/google.golang.org/grpc/client_test.go
@@ -2,82 +2,24 @@ package grpc
 
 import (
 	"context"
-	"encoding/json"
-	"log"
-	"net"
-	reflect "reflect"
 	"testing"
 
 	"github.com/golang/protobuf/ptypes/timestamp"
 	"github.com/stretchr/testify/assert"
-	grpcinternal "github.com/traceableai/goagent/otel/grpc/internal"
-	"github.com/traceableai/goagent/otel/internal"
+	grpcinternal "github.com/traceableai/goagent/instrumentation/google.golang.org/grpc/internal"
+	"github.com/traceableai/goagent/instrumentation/internal"
 	otel "go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc"
 	"go.opentelemetry.io/otel/api/global"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
-	"google.golang.org/grpc/test/bufconn"
 )
 
-var _ grpcinternal.PersonRegistryServer = server{}
-
-type server struct {
-	reply *grpcinternal.RegisterReply
-	err   error
-	*grpcinternal.UnimplementedPersonRegistryServer
-}
-
-func (s server) Register(_ context.Context, _ *grpcinternal.RegisterRequest) (*grpcinternal.RegisterReply, error) {
-	if s.reply == nil && s.err == nil {
-		log.Fatal("missing reply or error in server")
-	}
-
-	return s.reply, s.err
-}
-
-// createDialer creates a connection to be used as context dialer in GRPC
-// communication.
-func createDialer(s *grpc.Server) func(context.Context, string) (net.Conn, error) {
-	const bufSize = 1024 * 1024
-
-	listener := bufconn.Listen(bufSize)
-	conn := func(context.Context, string) (net.Conn, error) {
-		return listener.Dial()
-	}
-
-	go func() {
-		if err := s.Serve(listener); err != nil {
-			log.Fatalf("Server exited with error: %v", err)
-		}
-	}()
-
-	return conn
-}
-
-// jsonEqual compares the JSON from two strings.
-func jsonEqual(a, b string) (bool, error) {
-	var j, j2 interface{}
-	if err := json.Unmarshal([]byte(a), &j); err != nil {
-		return false, err
-	}
-	if err := json.Unmarshal([]byte(b), &j2); err != nil {
-		return false, err
-	}
-	return reflect.DeepEqual(j2, j), nil
-}
-
-func TestServerRegisterPersonSuccess(t *testing.T) {
+func TestClientRegisterPersonSuccess(t *testing.T) {
 	_, flusher := internal.InitTracer()
 
-	s := grpc.NewServer(
-		grpc.UnaryInterceptor(
-			WrapUnaryServerInterceptor(
-				otel.UnaryServerInterceptor(global.TraceProvider().Tracer("ai.traceable")),
-			),
-		),
-	)
+	s := grpc.NewServer()
 	defer s.Stop()
 
 	grpcinternal.RegisterPersonRegistryServer(s, &server{
@@ -92,6 +34,11 @@ func TestServerRegisterPersonSuccess(t *testing.T) {
 		"bufnet",
 		grpc.WithContextDialer(dialer),
 		grpc.WithInsecure(),
+		grpc.WithUnaryInterceptor(
+			WrapUnaryClientInterceptor(
+				otel.UnaryClientInterceptor(global.TraceProvider().Tracer("ai.traceable")),
+			),
+		),
 	)
 	if err != nil {
 		t.Fatalf("failed to dial bufnet: %v", err)
@@ -100,13 +47,16 @@ func TestServerRegisterPersonSuccess(t *testing.T) {
 
 	client := grpcinternal.NewPersonRegistryClient(conn)
 
-	ctx = metadata.NewOutgoingContext(ctx, metadata.Pairs("test_key", "test_value"))
-	_, err = client.Register(ctx, &grpcinternal.RegisterRequest{
-		Firstname: "Bugs",
-		Lastname:  "Bunny",
-		Birthdate: &timestamp.Timestamp{Seconds: 1},
-		Confirmed: false,
-	})
+	ctx = metadata.NewOutgoingContext(ctx, metadata.Pairs("test_key_1", "test_value_1"))
+	_, err = client.Register(
+		ctx,
+		&grpcinternal.RegisterRequest{
+			Firstname: "Bugs",
+			Lastname:  "Bunny",
+			Birthdate: &timestamp.Timestamp{Seconds: 1},
+			Confirmed: false,
+		},
+	)
 	if err != nil {
 		t.Fatalf("call to Register failed: %v", err)
 	}
@@ -121,7 +71,7 @@ func TestServerRegisterPersonSuccess(t *testing.T) {
 	assert.Equal(t, "grpc", attrs.Get("rpc.system").AsString())
 	assert.Equal(t, "helloworld.PersonRegistry", attrs.Get("rpc.service").AsString())
 	assert.Equal(t, "Register", attrs.Get("rpc.method").AsString())
-	assert.Equal(t, "test_value", attrs.Get("grpc.request.metadata.test_key").AsString())
+	assert.Equal(t, "test_value_1", attrs.Get("grpc.request.metadata.test_key_1").AsString())
 
 	expectedBody := "{\"firstname\":\"Bugs\",\"lastname\":\"Bunny\",\"birthdate\":\"1970-01-01T00:00:01Z\",\"confirmed\":false}"
 	if ok, err := jsonEqual(expectedBody, attrs.Get("grpc.request.body").AsString()); err == nil {
@@ -138,16 +88,10 @@ func TestServerRegisterPersonSuccess(t *testing.T) {
 	}
 }
 
-func TestServerRegisterPersonFails(t *testing.T) {
+func TestClientRegisterPersonFails(t *testing.T) {
 	_, flusher := internal.InitTracer()
 
-	s := grpc.NewServer(
-		grpc.UnaryInterceptor(
-			WrapUnaryServerInterceptor(
-				otel.UnaryServerInterceptor(global.TraceProvider().Tracer("ai.traceable")),
-			),
-		),
-	)
+	s := grpc.NewServer()
 	defer s.Stop()
 
 	expectedError := status.Error(codes.InvalidArgument, "invalid argument")
@@ -163,6 +107,11 @@ func TestServerRegisterPersonFails(t *testing.T) {
 		"bufnet",
 		grpc.WithContextDialer(dialer),
 		grpc.WithInsecure(),
+		grpc.WithUnaryInterceptor(
+			WrapUnaryClientInterceptor(
+				otel.UnaryClientInterceptor(global.TraceProvider().Tracer("ai.traceable")),
+			),
+		),
 	)
 	if err != nil {
 		t.Fatalf("failed to dial bufnet: %v", err)
@@ -189,16 +138,10 @@ func TestServerRegisterPersonFails(t *testing.T) {
 	assert.Equal(t, "invalid argument", span.StatusMessage)
 }
 
-func BenchmarkServerRequestResponseBodyMarshaling(b *testing.B) {
+func BenchmarkClientRequestResponseBodyMarshaling(b *testing.B) {
 	internal.InitTracer()
 
-	s := grpc.NewServer(
-		grpc.UnaryInterceptor(
-			WrapUnaryServerInterceptor(
-				otel.UnaryServerInterceptor(global.TraceProvider().Tracer("ai.traceable")),
-			),
-		),
-	)
+	s := grpc.NewServer()
 	defer s.Stop()
 
 	grpcinternal.RegisterPersonRegistryServer(s, &server{
@@ -213,6 +156,11 @@ func BenchmarkServerRequestResponseBodyMarshaling(b *testing.B) {
 		"bufnet",
 		grpc.WithContextDialer(dialer),
 		grpc.WithInsecure(),
+		grpc.WithUnaryInterceptor(
+			WrapUnaryClientInterceptor(
+				otel.UnaryClientInterceptor(global.TraceProvider().Tracer("ai.traceable")),
+			),
+		),
 	)
 	if err != nil {
 		b.Fatalf("failed to dial bufnet: %v", err)
@@ -236,12 +184,10 @@ func BenchmarkServerRequestResponseBodyMarshaling(b *testing.B) {
 	}
 }
 
-func BenchmarkServerRequestDefaultInterceptor(b *testing.B) {
+func BenchmarkClientRequestDefaultInterceptor(b *testing.B) {
 	tracer, _ := internal.InitTracer()
 
-	s := grpc.NewServer(
-		grpc.UnaryInterceptor(otel.UnaryServerInterceptor(tracer)),
-	)
+	s := grpc.NewServer()
 	defer s.Stop()
 
 	grpcinternal.RegisterPersonRegistryServer(s, &server{
@@ -256,6 +202,7 @@ func BenchmarkServerRequestDefaultInterceptor(b *testing.B) {
 		"bufnet",
 		grpc.WithContextDialer(dialer),
 		grpc.WithInsecure(),
+		grpc.WithUnaryInterceptor(otel.UnaryClientInterceptor(tracer)),
 	)
 	if err != nil {
 		b.Fatalf("failed to dial bufnet: %v", err)
