+++ instrumentation/google.golang.org/grpc/server_test.go
@@ -2,24 +2,82 @@ package grpc
 
 import (
 	"context"
+	"encoding/json"
+	"log"
+	"net"
+	reflect "reflect"
 	"testing"
 
 	"github.com/golang/protobuf/ptypes/timestamp"
 	"github.com/stretchr/testify/assert"
-	grpcinternal "github.com/traceableai/goagent/otel/grpc/internal"
-	"github.com/traceableai/goagent/otel/internal"
+	grpcinternal "github.com/traceableai/goagent/instrumentation/google.golang.org/grpc/internal"
+	"github.com/traceableai/goagent/instrumentation/internal"
 	otel "go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc"
 	"go.opentelemetry.io/otel/api/global"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
+	"google.golang.org/grpc/test/bufconn"
 )
 
-func TestClientRegisterPersonSuccess(t *testing.T) {
+var _ grpcinternal.PersonRegistryServer = server{}
+
+type server struct {
+	reply *grpcinternal.RegisterReply
+	err   error
+	*grpcinternal.UnimplementedPersonRegistryServer
+}
+
+func (s server) Register(_ context.Context, _ *grpcinternal.RegisterRequest) (*grpcinternal.RegisterReply, error) {
+	if s.reply == nil && s.err == nil {
+		log.Fatal("missing reply or error in server")
+	}
+
+	return s.reply, s.err
+}
+
+// createDialer creates a connection to be used as context dialer in GRPC
+// communication.
+func createDialer(s *grpc.Server) func(context.Context, string) (net.Conn, error) {
+	const bufSize = 1024 * 1024
+
+	listener := bufconn.Listen(bufSize)
+	conn := func(context.Context, string) (net.Conn, error) {
+		return listener.Dial()
+	}
+
+	go func() {
+		if err := s.Serve(listener); err != nil {
+			log.Fatalf("Server exited with error: %v", err)
+		}
+	}()
+
+	return conn
+}
+
+// jsonEqual compares the JSON from two strings.
+func jsonEqual(a, b string) (bool, error) {
+	var j, j2 interface{}
+	if err := json.Unmarshal([]byte(a), &j); err != nil {
+		return false, err
+	}
+	if err := json.Unmarshal([]byte(b), &j2); err != nil {
+		return false, err
+	}
+	return reflect.DeepEqual(j2, j), nil
+}
+
+func TestServerRegisterPersonSuccess(t *testing.T) {
 	_, flusher := internal.InitTracer()
 
-	s := grpc.NewServer()
+	s := grpc.NewServer(
+		grpc.UnaryInterceptor(
+			WrapUnaryServerInterceptor(
+				otel.UnaryServerInterceptor(global.TraceProvider().Tracer("ai.traceable")),
+			),
+		),
+	)
 	defer s.Stop()
 
 	grpcinternal.RegisterPersonRegistryServer(s, &server{
@@ -34,11 +92,6 @@ func TestClientRegisterPersonSuccess(t *testing.T) {
 		"bufnet",
 		grpc.WithContextDialer(dialer),
 		grpc.WithInsecure(),
-		grpc.WithUnaryInterceptor(
-			WrapUnaryClientInterceptor(
-				otel.UnaryClientInterceptor(global.TraceProvider().Tracer("ai.traceable")),
-			),
-		),
 	)
 	if err != nil {
 		t.Fatalf("failed to dial bufnet: %v", err)
@@ -47,16 +100,13 @@ func TestClientRegisterPersonSuccess(t *testing.T) {
 
 	client := grpcinternal.NewPersonRegistryClient(conn)
 
-	ctx = metadata.NewOutgoingContext(ctx, metadata.Pairs("test_key_1", "test_value_1"))
-	_, err = client.Register(
-		ctx,
-		&grpcinternal.RegisterRequest{
-			Firstname: "Bugs",
-			Lastname:  "Bunny",
-			Birthdate: &timestamp.Timestamp{Seconds: 1},
-			Confirmed: false,
-		},
-	)
+	ctx = metadata.NewOutgoingContext(ctx, metadata.Pairs("test_key", "test_value"))
+	_, err = client.Register(ctx, &grpcinternal.RegisterRequest{
+		Firstname: "Bugs",
+		Lastname:  "Bunny",
+		Birthdate: &timestamp.Timestamp{Seconds: 1},
+		Confirmed: false,
+	})
 	if err != nil {
 		t.Fatalf("call to Register failed: %v", err)
 	}
@@ -71,7 +121,7 @@ func TestClientRegisterPersonSuccess(t *testing.T) {
 	assert.Equal(t, "grpc", attrs.Get("rpc.system").AsString())
 	assert.Equal(t, "helloworld.PersonRegistry", attrs.Get("rpc.service").AsString())
 	assert.Equal(t, "Register", attrs.Get("rpc.method").AsString())
-	assert.Equal(t, "test_value_1", attrs.Get("grpc.request.metadata.test_key_1").AsString())
+	assert.Equal(t, "test_value", attrs.Get("grpc.request.metadata.test_key").AsString())
 
 	expectedBody := "{\"firstname\":\"Bugs\",\"lastname\":\"Bunny\",\"birthdate\":\"1970-01-01T00:00:01Z\",\"confirmed\":false}"
 	if ok, err := jsonEqual(expectedBody, attrs.Get("grpc.request.body").AsString()); err == nil {
@@ -88,10 +138,16 @@ func TestClientRegisterPersonSuccess(t *testing.T) {
 	}
 }
 
-func TestClientRegisterPersonFails(t *testing.T) {
+func TestServerRegisterPersonFails(t *testing.T) {
 	_, flusher := internal.InitTracer()
 
-	s := grpc.NewServer()
+	s := grpc.NewServer(
+		grpc.UnaryInterceptor(
+			WrapUnaryServerInterceptor(
+				otel.UnaryServerInterceptor(global.TraceProvider().Tracer("ai.traceable")),
+			),
+		),
+	)
 	defer s.Stop()
 
 	expectedError := status.Error(codes.InvalidArgument, "invalid argument")
@@ -107,11 +163,6 @@ func TestClientRegisterPersonFails(t *testing.T) {
 		"bufnet",
 		grpc.WithContextDialer(dialer),
 		grpc.WithInsecure(),
-		grpc.WithUnaryInterceptor(
-			WrapUnaryClientInterceptor(
-				otel.UnaryClientInterceptor(global.TraceProvider().Tracer("ai.traceable")),
-			),
-		),
 	)
 	if err != nil {
 		t.Fatalf("failed to dial bufnet: %v", err)
@@ -138,10 +189,16 @@ func TestClientRegisterPersonFails(t *testing.T) {
 	assert.Equal(t, "invalid argument", span.StatusMessage)
 }
 
-func BenchmarkClientRequestResponseBodyMarshaling(b *testing.B) {
+func BenchmarkServerRequestResponseBodyMarshaling(b *testing.B) {
 	internal.InitTracer()
 
-	s := grpc.NewServer()
+	s := grpc.NewServer(
+		grpc.UnaryInterceptor(
+			WrapUnaryServerInterceptor(
+				otel.UnaryServerInterceptor(global.TraceProvider().Tracer("ai.traceable")),
+			),
+		),
+	)
 	defer s.Stop()
 
 	grpcinternal.RegisterPersonRegistryServer(s, &server{
@@ -156,11 +213,6 @@ func BenchmarkClientRequestResponseBodyMarshaling(b *testing.B) {
 		"bufnet",
 		grpc.WithContextDialer(dialer),
 		grpc.WithInsecure(),
-		grpc.WithUnaryInterceptor(
-			WrapUnaryClientInterceptor(
-				otel.UnaryClientInterceptor(global.TraceProvider().Tracer("ai.traceable")),
-			),
-		),
 	)
 	if err != nil {
 		b.Fatalf("failed to dial bufnet: %v", err)
@@ -184,10 +236,12 @@ func BenchmarkClientRequestResponseBodyMarshaling(b *testing.B) {
 	}
 }
 
-func BenchmarkClientRequestDefaultInterceptor(b *testing.B) {
+func BenchmarkServerRequestDefaultInterceptor(b *testing.B) {
 	tracer, _ := internal.InitTracer()
 
-	s := grpc.NewServer()
+	s := grpc.NewServer(
+		grpc.UnaryInterceptor(otel.UnaryServerInterceptor(tracer)),
+	)
 	defer s.Stop()
 
 	grpcinternal.RegisterPersonRegistryServer(s, &server{
@@ -202,7 +256,6 @@ func BenchmarkClientRequestDefaultInterceptor(b *testing.B) {
 		"bufnet",
 		grpc.WithContextDialer(dialer),
 		grpc.WithInsecure(),
-		grpc.WithUnaryInterceptor(otel.UnaryClientInterceptor(tracer)),
 	)
 	if err != nil {
 		b.Fatalf("failed to dial bufnet: %v", err)
