+++ instrumentation/internal/spanattributes.go
