+++ instrumentation/internal/spanattributes_test.go
