+++ instrumentation/net/http/handler_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/traceableai/goagent/otel/internal"
+	"github.com/traceableai/goagent/instrumentation/internal"
 	otelhttp "go.opentelemetry.io/contrib/instrumentation/net/http"
 	apitrace "go.opentelemetry.io/otel/api/trace"
 )
